#300: Move tests to accessible package.
